requirements.txt
pandas
transformers
numpy
langchain
langchain_community
python-dotenv
huggingface_hub

dspy-ai

